backend/requirements-minimal.txt
# Ultra-minimal requirements for Render free tier (512MB RAM)
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
boto3==1.34.0

# Lightweight ML dependencies
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# PDF processing
PyPDF2==3.0.1

# HTTP requests
requests==2.31.0
